command/generate: Parse record-count as a u64
The argument parser never allowed negative numbers, so this change still
includes the entire previous input set.

+++ src/commands/generate.rs
@@ -8,7 +8,7 @@ pub fn generate(matches: &ArgMatches<'_>) -> anyhow::Result<()> {
     let r1_dst = matches.value_of("r1-dst").unwrap();
     let r2_dst = matches.value_of("r2-dst").unwrap();
 
-    let record_count = value_t!(matches, "record-count", i32).unwrap_or_else(|e| e.exit());
+    let record_count = value_t!(matches, "record-count", u64).unwrap_or_else(|e| e.exit());
 
     info!("fq-generate start");
